add confirm delete dialog

+++ src/app/pages/dashboard/setting/pages/home-slider-banner/home-slider-banner.component.ts
@@ -1,9 +1,11 @@
 import { Component, EventEmitter, OnDestroy, OnInit, Output } from '@angular/core';
 import {FormControl, FormGroup, Validators } from '@angular/forms';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Editor, Toolbar } from 'ngx-editor';
 import { ToastrService } from 'ngx-toastr';
 import { SliderTypes } from 'src/app/shared/Enums/enums';
 import { environment } from 'src/environments/environment';
+import { ConfirmationDialogComponent } from '../../../confirmation-dialog/confirmation-dialog.component';
 import jsonDoc from '../../models/doc';
 import { ISlider, ISliderAttachment } from '../../models/slider.interface';
 import { SliderService } from '../../services/slider.service';
@@ -37,7 +39,7 @@ export class HomeSliderBannerComponent implements OnInit{
 
 
 
-  constructor(private sliderService: SliderService,private toastr: ToastrService) {
+  constructor(private sliderService: SliderService,private toastr: ToastrService,private modalService: NgbModal) {
   }
 
   ngOnInit(): void {
@@ -166,17 +168,41 @@ export class HomeSliderBannerComponent implements OnInit{
   }
 
   DeleteRow(attachmentId:number){
-    this.sliderService.deleteSliderAttachment(attachmentId).subscribe(r => {
-      if (!r.isError) {
-        this.toastr.success("Successfully Deleted")
-        this.getAllSliderAttatchments();
-        this.ngOnInit();
-        
+
+    this.modalService.open(ConfirmationDialogComponent, { size: 'sm' }).closed.subscribe(res => {
+      if (res) {
+        this.sliderService.deleteSliderAttachment(attachmentId).subscribe(r => {
+          if (!r.isError) {
+            this.toastr.success("Successfully Deleted")
+            this.getAllSliderAttatchments();
+            this.ngOnInit();
+            
+          }
+          else{
+            this.toastr.error("Failed Deleted")
+          }
+        });
       }
-      else{
-        this.toastr.error("Failed Deleted")
+      else {
+        // this.note.warn(':: test')
       }
     });
+
+
+
+
+
+    // this.sliderService.deleteSliderAttachment(attachmentId).subscribe(r => {
+    //   if (!r.isError) {
+    //     this.toastr.success("Successfully Deleted")
+    //     this.getAllSliderAttatchments();
+    //     this.ngOnInit();
+        
+    //   }
+    //   else{
+    //     this.toastr.error("Failed Deleted")
+    //   }
+    // });
 
   }
